pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "balr"
version = "0.1.0"
description = "Binary-Attribute-based Likelihood Ratio estimation for speaker recognition"
readme = "README.md"
license = { text = "MIT" }
authors = [
    {name = "Heiko Agnoli", email = "heiko.agnoli@inria.fr"},
    {name = "Jean-François Bonastre", email = "jean-francois.bonastre@inria.fr"},
    {name = "Raphael Duroselle", email = "raphael.duroselle@inria.fr"},
    {name = "Yosra Jelassi", email = "yosra.jelassi@inria.fr"},
    {name = "Jonas Renault", email = "jonas.renault@inria.fr"},
    {name = "Imen Ben Amor", email = "imen.ben-amor@univ-avignon.fr"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.29.2",
    "hydra-core>=1.3.2",
    "psutil>=7.0.0",
    "pytorch-metric-learning>=2.8.1",
    "seaborn>=0.13.2",
    "soundfile>=0.13.1",
    "speechbrain>=1.0.2",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "tqdm>=4.67.1",
    "typer>=0.15.2",
    "wespeaker",
]

[project.urls]
"Homepage" = "https://github.com/inria-defense/balr#readme"
"Repository" = "https://github.com/inria-defense/balr"

[project.scripts]
balr = "balr.cli.main:app"

[project.optional-dependencies]
docs = [
    "mkdocs-material>=9.6.11",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "jupyterlab>=4.3.5",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "ruff>=0.9.9",
]

[tool.black]
line_length = 90
target_version = ['py312']

[tool.ruff]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
]
src = ["balr", "tests"]
line-length = 90

[tool.ruff.lint]
ignore = []
select = [
    "E",
    "F",
    "W",
    "I001",
]

[tool.ruff.lint.isort]
known-first-party = ["balr"]

[tool.mypy]
files = ["balr", "tests"]
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    'ignore::DeprecationWarning',
]

[tool.uv.sources]
wespeaker = { git = "https://github.com/wenet-e2e/wespeaker.git" }
